rename Tabel MainTable; make form url an env
